Add multiple conditions for inconsistent recipes
Use formatted string for recipe image with ID to anticipate "null" image value. If ingredient clean name is null, skip the ingredient -- example was crockpot liner.  If ingredient, unit, quantity combination is repeated, skip it. If the ingredient & unit are the same, but differ in quantity, add to previous quantity. Ex: 1 egg, 4 eggs. Multiple print statements used in testing.